refctor brctl into pkg and cmd component
